split tests folder from xclim package, move testing helpers to xclim/testing, adjust pytest options and calls
