+++ setup.cfg
@@ -3,13 +3,13 @@ current_version = 0.42.0
 commit = True
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+).(?P<patch>\d+)(\-(?P<release>[a-z]+))?
-serialize = 
+serialize =
 	{major}.{minor}.{patch}-{release}
 	{major}.{minor}.{patch}
 
 [bumpversion:part:release]
 optional_value = gamma
-values = 
+values =
 	beta
 	gamma
 
@@ -18,15 +18,15 @@ search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
 
 [flake8]
-exclude = 
+exclude =
 	.git,
 	docs,
 	build,
 	.eggs,
-	xclim/testing/tests
+	tests
 max-line-length = 88
 max-complexity = 12
-ignore = 
+ignore =
 	C901
 	E203
 	E231
@@ -36,12 +36,12 @@ ignore =
 	F403
 	W503
 	W504
-per-file-ignores = 
+per-file-ignores =
 	tests/*:E402
-rst-directives = 
+rst-directives =
 	bibliography
 	autolink-skip
-rst-roles = 
+rst-roles =
 	doc,
 	mod,
 	py:attr,
@@ -60,7 +60,7 @@ rst-roles =
 	cite:p
 	cite:t
 	cite:ts
-extend-ignore = 
+extend-ignore =
 	RST399,
 	RST201,
 	RST203,
@@ -70,8 +70,8 @@ extend-ignore =
 
 [pycodestyle]
 count = False
-exclude = xclim/testing/tests
-ignore = 
+exclude = tests
+ignore =
 	E226,
 	E402,
 	E501,
