+++ tests/test_generic_indicators.py
