+++ tests/test_indicators.py
