+++ tests/test_options.py
