+++ tests/test_sdba/test_detrending.py
