+++ tests/test_stats.py
