--- xclim/testing/tests/__init__.py
@@ -1,5 +1,0 @@
-# noqa: D104
-from __future__ import annotations
-
-from .data import TD
-from .test_sdba import *
