--- xclim/testing/tests/test_sdba/__init__.py
@@ -1,5 +1,0 @@
-# noqa: D104
-from __future__ import annotations
-
-from .diagnostics import *
-from .utils import *
